Text+To+SQL+LLM/text+to+sql+rag+llm/requirements.txt
streamlit
pandas
sqlalchemy
pymysql
mysql-connector-python
requests
pydantic-settings
langchain-core
langchain_ollama
langchain_classic
langchain_chroma
langchain_community
sentence_transformers
cryptography